+++ packages/storybook/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.9.0](https://github.com/ab-devtools/uilibrary/compare/@ab.uitools/storybook@1.8.8...@ab.uitools/storybook@1.9.0) (2026-06-16)
+
+### Features
+
+- add onChipClick prop to support global click events and update stories accordingly ([92eedc4](https://github.com/ab-devtools/uilibrary/commit/92eedc40409c7dc234f4b090b4222ecb943b431f))
+
+## [1.8.8](https://github.com/ab-devtools/uilibrary/compare/@ab.uitools/storybook@1.8.7...@ab.uitools/storybook@1.8.8) (2026-06-15)
+
+**Note:** Version bump only for package @ab.uitools/storybook
+
 ## [1.8.7](https://github.com/ab-devtools/uilibrary/compare/@ab.uitools/storybook@1.8.6...@ab.uitools/storybook@1.8.7) (2026-06-12)
 
 **Note:** Version bump only for package @ab.uitools/storybook