+++ packages/ui-library/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.20.0](https://github.com/ab-ui-tools/ab-ui-tools/compare/@ab.uitools/ui-library@1.19.6...@ab.uitools/ui-library@1.20.0) (2026-06-16)
+
+### Features
+
+- add onChipClick prop to support global click events and update stories accordingly ([92eedc4](https://github.com/ab-ui-tools/ab-ui-tools/commit/92eedc40409c7dc234f4b090b4222ecb943b431f))
+
+## [1.19.6](https://github.com/ab-ui-tools/ab-ui-tools/compare/@ab.uitools/ui-library@1.19.5...@ab.uitools/ui-library@1.19.6) (2026-06-15)
+
+### Bug Fixes
+
+- one time password ([3fe2e97](https://github.com/ab-ui-tools/ab-ui-tools/commit/3fe2e9742e040dc1781239a131a7f025672bc0f5))
+- otp ([6008bcd](https://github.com/ab-ui-tools/ab-ui-tools/commit/6008bcd6df8a16af66e99bfcc1717d83ae83a463))
+- otp ([96f62c1](https://github.com/ab-ui-tools/ab-ui-tools/commit/96f62c1a7952240967b6b4a10b1aecda061e3867))
+
 ## [1.19.5](https://github.com/ab-ui-tools/ab-ui-tools/compare/@ab.uitools/ui-library@1.19.4...@ab.uitools/ui-library@1.19.5) (2026-06-12)
 
 ### Bug Fixes